namespaces, header files and tests
