+++ library/utils.h
@@ -1,7 +1,10 @@
+#pragma once
+
 #include <stdio.h>
 
 #pragma once
 
+namespace PropertyModel {
 namespace Library {
 
 template<size_t Index>
@@ -12,4 +15,5 @@ struct Value;
 
 template<size_t Index>
 struct Output;
-} // namespace Library
+}; // namespace Library
+}; // namespace PropertyModel
